requirements.txt
fastapi==0.115.2
uvicorn==0.31.0
faiss-cpu==1.8.0
sentence-transformers==3.1.1
pandas==2.2.2
httpx==0.27.0
litellm==1.48.0
python-dotenv==1.0.1
